requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
redis
werkzeug
python-dotenv